BUGFIX: make it compile with -Wall -Werror

+++ tests/headers_test.cc
@@ -9,7 +9,7 @@ TEST(headers_test, accept) {
 
     {
         const auto& media = a1.media();
-        ASSERT_EQ(media.size(), 1);
+        ASSERT_EQ(media.size(), size_t(1));
 
         const auto& mime = media[0];
         ASSERT_EQ(mime, MIME(Audio, Star));
@@ -21,7 +21,7 @@ TEST(headers_test, accept) {
 
     {
         const auto& media = a2.media();
-        ASSERT_EQ(media.size(), 4);
+        ASSERT_EQ(media.size(), size_t(4));
 
         const auto &m1 = media[0];
         ASSERT_EQ(m1, MIME(Text, Star));
@@ -41,7 +41,7 @@ TEST(headers_test, accept) {
 
     {
         const auto& media = a3.media();
-        ASSERT_EQ(media.size(), 5);
+        ASSERT_EQ(media.size(), size_t(5));
 
         ASSERT_EQ(media[0], MIME(Text, Star));
         ASSERT_EQ(media[0].q().getOrElse(Mime::Q(0)), Mime::Q(30));
@@ -95,7 +95,7 @@ TEST(headers_test, cache_control) {
         cc.parse(str);
 
         auto directives = cc.directives();
-        ASSERT_EQ(directives.size(), 1);
+        ASSERT_EQ(directives.size(), size_t(1));
         ASSERT_EQ(directives[0].directive(), expected);
     };
 
@@ -105,7 +105,7 @@ TEST(headers_test, cache_control) {
         cc.parse(str);
 
         auto directives = cc.directives();
-        ASSERT_EQ(directives.size(), 1);
+        ASSERT_EQ(directives.size(), size_t(1));
 
         ASSERT_EQ(directives[0].directive(), expected);
         ASSERT_EQ(directives[0].delta(), std::chrono::seconds(delta));
@@ -125,7 +125,7 @@ TEST(headers_test, cache_control) {
     Header::CacheControl cc1;
     cc1.parse("private, max-age=600");
     auto d1 = cc1.directives();
-    ASSERT_EQ(d1.size(), 2);
+    ASSERT_EQ(d1.size(), size_t(2));
     ASSERT_EQ(d1[0].directive(), CacheDirective::Private);
     ASSERT_EQ(d1[1].directive(), CacheDirective::MaxAge);
     ASSERT_EQ(d1[1].delta(), std::chrono::seconds(600));
@@ -133,7 +133,7 @@ TEST(headers_test, cache_control) {
     Header::CacheControl cc2;
     cc2.parse("public, s-maxage=200, proxy-revalidate");
     auto d2 = cc2.directives();
-    ASSERT_EQ(d2.size(), 3);
+    ASSERT_EQ(d2.size(), size_t(3));
     ASSERT_EQ(d2[0].directive(), CacheDirective::Public);
     ASSERT_EQ(d2[1].directive(), CacheDirective::SMaxAge);
     ASSERT_EQ(d2[1].delta(), std::chrono::seconds(200));
@@ -164,7 +164,7 @@ TEST(headers_test, content_length) {
     Header::ContentLength cl;
 
     cl.parse("3495");
-    ASSERT_EQ(cl.value(), 3495);
+    ASSERT_EQ(cl.value(), size_t(3495));
 }
 
 TEST(headers_test, connection) {
